+++ src/io/openems/impl/protocol/modbus/ModbusChannel.java
@@ -1,19 +1,18 @@
 package io.openems.impl.protocol.modbus;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 import io.openems.api.channel.Channel;
 import io.openems.api.device.nature.DeviceNature;
 
 public class ModbusChannel extends Channel {
-	public ModbusChannel(DeviceNature nature, String unit, BigInteger minValue, BigInteger maxValue,
-			BigInteger multiplier, BigInteger delta, Map<BigInteger, String> labels) {
+	public ModbusChannel(DeviceNature nature, String unit, Long minValue, Long maxValue, Long multiplier, Long delta,
+			Map<Long, String> labels) {
 		super(nature, unit, minValue, maxValue, multiplier, delta, labels);
 	}
 
 	@Override
-	protected void updateValue(BigInteger value) {
+	protected void updateValue(Long value) {
 		super.updateValue(value);
 	}
 }
