+++ src/io/openems/impl/protocol/modbus/WriteableModbusChannel.java
@@ -1,6 +1,5 @@
 package io.openems.impl.protocol.modbus;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 import io.openems.api.channel.Channel;
@@ -9,15 +8,15 @@ import io.openems.api.device.nature.DeviceNature;
 
 public class WriteableModbusChannel extends WriteableChannel {
 
-	public WriteableModbusChannel(DeviceNature nature, String unit, BigInteger minValue, BigInteger maxValue,
-			BigInteger multiplier, BigInteger delta, Map<BigInteger, String> labels, BigInteger minWriteValue,
-			Channel minWriteValueChannel, BigInteger maxWriteValue, Channel maxWriteValueChannel) {
+	public WriteableModbusChannel(DeviceNature nature, String unit, Long minValue, Long maxValue, Long multiplier,
+			Long delta, Map<Long, String> labels, Long minWriteValue, Channel minWriteValueChannel, Long maxWriteValue,
+			Channel maxWriteValueChannel) {
 		super(nature, unit, minValue, maxValue, multiplier, delta, labels, minWriteValue, minWriteValueChannel,
 				maxWriteValue, maxWriteValueChannel);
 	}
 
 	@Override
-	protected void updateValue(BigInteger value) {
+	protected void updateValue(Long value) {
 		super.updateValue(value);
 	}
 }
