+++ src/io/openems/impl/protocol/modbus/internal/DoublewordElement.java
@@ -1,7 +1,5 @@
 package io.openems.impl.protocol.modbus.internal;
 
-import java.math.BigInteger;
-
 import org.eclipse.jdt.annotation.NonNull;
 
 import com.ghgande.j2mod.modbus.procimg.Register;
@@ -22,5 +20,5 @@ public interface DoublewordElement {
 	 * @param value
 	 * @return
 	 */
-	public Register[] toRegisters(@NonNull BigInteger value);
+	public Register[] toRegisters(@NonNull Long value);
 }
