+++ src/io/openems/impl/protocol/modbus/internal/UnsignedWordElement.java
@@ -17,7 +17,6 @@
  */
 package io.openems.impl.protocol.modbus.internal;
 
-import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 
@@ -47,11 +46,11 @@ public class UnsignedWordElement extends ModbusElement implements WordElement {
 		ByteBuffer buff = ByteBuffer.allocate(2).order(byteOrder);
 		buff.put(register.toBytes());
 		int shortValue = Short.toUnsignedInt(buff.getShort(0));
-		setValue(BigInteger.valueOf(shortValue));
+		setValue(Long.valueOf(shortValue));
 	}
 
 	@Override
-	public Register toRegister(@NonNull BigInteger value) {
+	public Register toRegister(@NonNull Long value) {
 		byte[] b = ByteBuffer.allocate(2).order(byteOrder).putShort(value.shortValue()).array();
 		return new SimpleRegister(b[0], b[1]);
 	}
